Added homebrew cask updater
Homebrew casks can't be updated as easily as normal brews... today it's not fully supported. This utility makes it easier to see the version installed. However, some versions aren't correct... if they self-update (VSCode / Slack / etc) the versions homebrew things are installed and the versions that are installed may conflict so better to check what needs to be updated.